split warp-apply
